+++ backend/wms/schemas/inventory.py
@@ -13,7 +13,7 @@ class LotOut(BaseModel):
     location_code: str | None
     location_is_overflow: bool
     location_is_qa_hold: bool
-    quantity: int
+    quantity: float  # SCO-143: base UoM, decimal-capable
     qa_hold: bool
     received_at: datetime
     expires_at: date | None
@@ -35,19 +35,19 @@ class SKUDetailOut(BaseModel):
     sku_code: str
     description: str
     uom: str
-    reorder_point: int
-    safety_stock: int
-    on_hand_total: int
-    available: int  # excludes qa_hold + expired
-    qa_hold_qty: int
-    expired_qty: int
+    reorder_point: float  # SCO-143
+    safety_stock: float  # SCO-143
+    on_hand_total: float  # SCO-143
+    available: float  # excludes qa_hold + expired
+    qa_hold_qty: float
+    expired_qty: float
     lot_count: int
 
 
 class InventoryKPIs(BaseModel):
-    total_on_hand: int
-    available: int
-    qa_hold_qty: int
+    total_on_hand: float  # SCO-143
+    available: float
+    qa_hold_qty: float
     qa_hold_lots: int
     slow_movers: int  # SKUs with zero outbound movement (placeholder until reports module)
     skus_below_safety: int
@@ -57,21 +57,21 @@ class InventoryKPIs(BaseModel):
 
 class AdjustRequest(BaseModel):
     lot_id: int
-    delta: int  # may be negative
+    delta: float  # SCO-143: may be negative; decimal-capable
     reason: str = Field(min_length=1, max_length=255)
 
 
 class AdjustOut(BaseModel):
     lot_id: int
-    was: int
-    now: int
-    delta: int
+    was: float  # SCO-143
+    now: float
+    delta: float
 
 
 class BelowSafetyRow(BaseModel):
     sku_code: str
     description: str
-    available: int
-    reorder_point: int
-    safety_stock: int
-    shortfall: int  # safety_stock - available, always positive
+    available: float  # SCO-143
+    reorder_point: float
+    safety_stock: float
+    shortfall: float  # safety_stock - available, always positive
